warning fixes for viewer directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23638 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/viewer/main/ActionGroupDescription.C
@@ -705,7 +705,7 @@ ActionGroupDescription::GetAction(int index) const
 {
     ViewerRPC::ViewerRPCType retval = ViewerRPC::MaxRPC;
 
-    if(index >= 0 && index < actions.size())
+    if(index >= 0 && index < (int)actions.size())
     {
         ViewerRPC::ViewerRPCType_FromString(actions[index], retval);
     }
